catalog: check if files exit before deleting them, #TASK-7047, #TASK-6442

+++ opencga-catalog/src/main/java/org/opencb/opencga/catalog/utils/ResourceManager.java
@@ -388,8 +388,14 @@ public class ResourceManager  {
             throw new ToolException("Error unzipping resource file '" + filename + FOR_ANALYSIS_MSG + analysisId + "'. Check log files: "
                     + stdoutPath + ", " + stderrPath, e);
         }
-        Files.delete(stdoutPath);
-        Files.delete(stderrPath);
+        
+        // Deleting stdout and stderr after unzipping
+        if (Files.exists(stdoutPath)) {
+            Files.delete(stdoutPath);
+        }
+        if (Files.exists(stderrPath)) {
+            Files.delete(stderrPath);
+        }
     }
 
     //-------------------------------------------------------------------------
